gossipd: use u64 for `htlc_minimum_msat` field
It's u64 in the spec, so we should use u64 too.

+++ gossipd/routing.h
@@ -27,7 +27,7 @@ struct half_chan {
 	s64 last_timestamp;
 
 	/* Minimum number of msatoshi in an HTLC */
-	u32 htlc_minimum_msat;
+	u64 htlc_minimum_msat;
 
 	/* Maximum number of msatoshis in an HTLC */
 	u64 htlc_maximum_msat;
